requirements.txt
python-dotenv
aiohttp